refactor(daemon): IRepairOperation, one file per repair operation
RepairService.cpp carried the thirteen on-demand repair operations as
sibling private methods (2,150 lines) behind a 13-way switch, a second
enum->name switch, and a third name->code chain. Adding an operation
meant touching the enum, three tables, the header, and a 3,400-line
file. Each operation only ever read the service's context bundle, its
config, the state component, and the running flag, so the seam was
already there.

repair_operation.h now declares IRepairOperation (name, wire code,
run, and an async variant that defaults to run) plus OperationEnv, the
bundle of those four things that RepairService builds per call. Each
operation lives in src/daemon/components/repair/ops/<name>.cpp as a
free function with the original body, moved verbatim (whitespace-
normalized comparison of all seventeen moved functions is identical
after the mechanical substitutions: ctx_/cfg_/state_/running_ ->
env.*, companion helpers gain the env parameter, and the enclosing
namespace forced ::yams::repair:: qualification), wrapped in a small
operation class. The helpers two or more operations share
(submitRepairStatusUpdate, queueWithBackoff, the MIME redetection
trio, the metadata text cap) move to repair_operation_support.h/.cpp;
the service keeps its own users of them through using-declarations.
The registry (repair_operation_registry.cpp) owns the thirteen
instances in wire-code order.

RepairService::runRepairOperation looks the operation up by code and
runs it; executeRepairAsync runs every non-foreground operation
through runAsync, which for the eleven operations without a coroutine
body is exactly the synchronous call that used to be inlined, and for
stuck-documents and embeddings is the same coroutine as before. The
enum->name switch becomes a lookup on the registry, so the metric key
table in metric_keys.h is now the only place an operation is named.

Tests: repair_operation_registry_catch2_test.cpp pins one operation
per wire code, canonical names, wire-code order, and unknown codes.
The repair_service, repair_submodule, and daemon_protocol repair
suites pass unchanged; a full debug compile is clean.

RepairService.cpp: 3,431 -> 1,090 lines; ops/: 13 files, 3,312 lines.

The enum-to-wire-code mapping (enumerator + 1) is pinned by a
static_assert against the table's size and a comment on the enum, since
the exhaustive switch that used to catch a missing case is gone.
