+++ src/daemon/components/RepairService.cpp
@@ -1,5 +1,7 @@
+#include "repair/repair_operation_support.h"
 #include <yams/daemon/components/db_salvage.h>
 #include <yams/daemon/components/MetadataWriteFacade.h>
+#include <yams/daemon/components/repair/repair_operation.h>
 #include <yams/daemon/components/RepairService.h>
 #include <yams/daemon/components/WriteCoordinator.h>
 
@@ -58,26 +60,9 @@ namespace yams::daemon {
 
 namespace {
 
-constexpr size_t kMaxTextToPersistInMetadataBytes =
-    16ULL * 1024ULL * 1024ULL; // 16 MiB (best-effort)
-
-template <typename Meta>
-inline void submitRepairStatusUpdate(const RepairServiceContext& ctx, Meta& metaRepo,
-                                     std::vector<std::string> hashes,
-                                     yams::metadata::RepairStatus status, std::string_view source) {
-    if (hashes.empty())
-        return;
-    if (auto* coord = ctx.getWriteCoordinator ? ctx.getWriteCoordinator() : nullptr) {
-        auto wb = std::make_unique<WriteBatch>();
-        wb->source.assign(source.data(), source.size());
-        wb->ops.emplace_back(UpdateRepairStatusOp{hashes, status});
-        coord->enqueue(std::unique_ptr<WriteBatch>(wb.release()));
-        return;
-    }
-    if (metaRepo) {
-        (void)metaRepo->batchUpdateDocumentRepairStatuses(hashes, status);
-    }
-}
+using repair::kMaxTextToPersistInMetadataBytes;
+using repair::queueWithBackoff;
+using repair::submitRepairStatusUpdate;
 
 uint64_t steadyNowMillis() {
     return static_cast<uint64_t>(std::chrono::duration_cast<std::chrono::milliseconds>(
@@ -92,98 +77,6 @@ uint64_t repairOperationCode(std::string_view operation) {
 // Check the shared vector compatibility policy used by search and initialization.
 bool vectorsDisabledByEnv() {
     return !yams::config::resolve_vector_environment().enabled;
-}
-
-std::string normalizedRepairExtension(const metadata::DocumentInfo& doc) {
-    std::string extension = doc.fileExtension;
-    if (extension.empty()) {
-        auto pos = doc.fileName.rfind('.');
-        if (pos != std::string::npos)
-            extension = doc.fileName.substr(pos);
-    }
-    if (extension.empty())
-        return {};
-    if (extension.front() != '.')
-        extension.insert(extension.begin(), '.');
-    std::transform(extension.begin(), extension.end(), extension.begin(),
-                   [](unsigned char c) { return static_cast<char>(std::tolower(c)); });
-    return extension;
-}
-
-bool shouldRedetectMime(const metadata::DocumentInfo& doc) {
-    if (doc.mimeType.empty() || doc.mimeType == "application/octet-stream")
-        return true;
-
-    if (doc.mimeType != "text/plain")
-        return false;
-
-    const auto extension = normalizedRepairExtension(doc);
-    if (extension.empty())
-        return true;
-
-    const auto hintedMime = detection::FileTypeDetector::getMimeTypeFromExtension(extension);
-    return !hintedMime.empty() && hintedMime != "text/plain";
-}
-
-std::string bestEffortMimeForDocument(const metadata::DocumentInfo& doc,
-                                      const std::shared_ptr<api::IContentStore>& store) {
-    try {
-        (void)detection::FileTypeDetector::initializeWithMagicNumbers();
-        auto& detector = detection::FileTypeDetector::instance();
-        const auto extension = normalizedRepairExtension(doc);
-        const auto hintedMime =
-            extension.empty() ? std::string{}
-                              : detection::FileTypeDetector::getMimeTypeFromExtension(extension);
-
-        if (store) {
-            // MIME repair only needs the leading magic bytes. Avoid materializing full objects
-            // here: `repair --all` can scan large corpora and full reads can stall streaming
-            // progress or exhaust daemon memory, surfacing to the CLI as an IPC EOF if the daemon
-            // is killed.
-            constexpr std::size_t kMimeSniffBytes = 8192;
-            auto bytesResult = store->retrieveBytesPrefix(doc.sha256Hash, kMimeSniffBytes);
-            if (bytesResult && !bytesResult.value().empty()) {
-                const auto& bytes = bytesResult.value();
-                auto detected = detector.detectFromBuffer(
-                    std::span<const std::byte>(bytes.data(), bytes.size()));
-                if (detected) {
-                    std::string mime = detected.value().mimeType;
-                    if (!hintedMime.empty() && hintedMime != "application/octet-stream") {
-                        const auto detectedType = detector.getFileTypeCategory(mime);
-                        const auto hintedType = detector.getFileTypeCategory(hintedMime);
-                        if (mime.empty() || mime == "application/octet-stream" ||
-                            (mime == "text/plain" && hintedMime != "text/plain") ||
-                            (detectedType == "executable" && hintedType == "executable" &&
-                             mime != hintedMime)) {
-                            mime = hintedMime;
-                        }
-                    }
-                    if (!mime.empty())
-                        return mime;
-                }
-            }
-        }
-
-        if (!doc.filePath.empty() && std::filesystem::exists(doc.filePath)) {
-            if (auto detected = detector.detectFromFile(doc.filePath)) {
-                const auto& mime = detected.value().mimeType;
-                if (!mime.empty())
-                    return mime;
-            }
-        }
-
-        if (!extension.empty()) {
-            const auto mime = hintedMime;
-            if (!mime.empty())
-                return mime;
-        }
-    } catch (const std::exception& e) {
-        spdlog::debug("RepairService: MIME sniff failed for {}: {}", doc.sha256Hash, e.what());
-    } catch (...) {
-        spdlog::debug("RepairService: MIME sniff failed for {}: unknown exception", doc.sha256Hash);
-    }
-
-    return "application/octet-stream";
 }
 
 // Shared thread pool for RepairService background coroutines
@@ -243,45 +136,6 @@ bool canExtractDocument(
         }
     }
     return false;
-}
-
-// Template helper: queue job with exponential backoff when full
-template <typename JobT>
-boost::asio::awaitable<bool>
-queueWithBackoff(std::shared_ptr<SpscQueue<JobT>> queue, JobT&& job,
-                 boost::asio::steady_timer& timer, std::atomic<bool>& running,
-                 const std::string& jobTypeName, size_t jobSize, int initialDelayMs = 50,
-                 int maxDelayMs = 1000, std::function<bool()> shouldStop = {},
-                 std::function<void()> onQueueFull = {}) {
-    const auto queueStart = std::chrono::steady_clock::now();
-    uint64_t totalWaitMs = 0;
-    int retries = 0;
-    while (!queue->try_push(std::forward<JobT>(job))) {
-        if (!running.load(std::memory_order_relaxed) || (shouldStop && shouldStop()))
-            co_return false;
-        if (retries == 0 && onQueueFull)
-            onQueueFull();
-        const int cappedRetries = std::min(retries, 5);
-        const int delayMs = std::min(initialDelayMs * (1 << cappedRetries), maxDelayMs);
-        totalWaitMs += static_cast<uint64_t>(delayMs);
-        const auto queued = queue ? queue->size_approx() : 0;
-        const auto capacity = queue ? queue->capacity() : 0;
-        spdlog::warn("RepairService: {} queue full, retry={} sleep_ms={} queued={} capacity={} "
-                     "batch_size={}",
-                     jobTypeName, retries + 1, delayMs, queued, capacity, jobSize);
-        timer.expires_after(std::chrono::milliseconds(delayMs));
-        co_await timer.async_wait(boost::asio::use_awaitable);
-        retries++;
-    }
-    if (retries > 0) {
-        const auto pushLatencyMs = std::chrono::duration_cast<std::chrono::milliseconds>(
-                                       std::chrono::steady_clock::now() - queueStart)
-                                       .count();
-        spdlog::info("RepairService: queued {} job after {} retries (batch_size={} wait_ms={} "
-                     "push_latency_ms={})",
-                     jobTypeName, retries, jobSize, totalWaitMs, pushLatencyMs);
-    }
-    co_return true;
 }
 
 } // namespace
@@ -1125,75 +979,27 @@ RepairService::buildRepairPlan(const RepairRequest& request) {
 }
 
 std::string_view RepairService::repairOperationName(OnDemandRepairOperation operation) noexcept {
-    switch (operation) {
-        case OnDemandRepairOperation::StuckDocuments:
-            return "stuck_docs";
-        case OnDemandRepairOperation::Orphans:
-            return "orphans";
-        case OnDemandRepairOperation::Mime:
-            return "mime";
-        case OnDemandRepairOperation::Downloads:
-            return "downloads";
-        case OnDemandRepairOperation::PathTree:
-            return "path_tree";
-        case OnDemandRepairOperation::Dedupe:
-            return "dedupe";
-        case OnDemandRepairOperation::Chunks:
-            return "chunks";
-        case OnDemandRepairOperation::BlockReferences:
-            return "block_refs";
-        case OnDemandRepairOperation::Graph:
-            return "graph";
-        case OnDemandRepairOperation::Fts5:
-            return "fts5";
-        case OnDemandRepairOperation::Embeddings:
-            return "embeddings";
-        case OnDemandRepairOperation::Topology:
-            return "topology";
-        case OnDemandRepairOperation::Optimize:
-            return "optimize";
-    }
-    return "unknown";
+    const auto* op = repair::repairOperationForCode(static_cast<std::uint64_t>(operation) + 1);
+    return op ? op->name() : std::string_view{"unknown"};
 }
 
 RepairOperationResult RepairService::runRepairOperation(OnDemandRepairOperation operation,
                                                         const RepairRequest& request,
                                                         const ProgressFn& progress,
                                                         std::atomic<bool>* cancelRequested) {
-    switch (operation) {
-        case OnDemandRepairOperation::StuckDocuments:
-            return recoverStuckDocuments(request, progress);
-        case OnDemandRepairOperation::Orphans:
-            return cleanOrphanedMetadata(request.dryRun, request.verbose, request.removeCorrupt,
-                                         progress);
-        case OnDemandRepairOperation::Mime:
-            return repairMimeTypes(request.dryRun, request.verbose, progress);
-        case OnDemandRepairOperation::Downloads:
-            return repairDownloads(request.dryRun, request.verbose, progress);
-        case OnDemandRepairOperation::PathTree:
-            return rebuildPathTree(request.dryRun, request.verbose, progress);
-        case OnDemandRepairOperation::Dedupe:
-            return applySemanticDedupe(request, progress);
-        case OnDemandRepairOperation::Chunks:
-            return cleanOrphanedChunks(request.dryRun, request.verbose, progress);
-        case OnDemandRepairOperation::BlockReferences:
-            return repairBlockReferences(request.dryRun, request.verbose, progress);
-        case OnDemandRepairOperation::Graph:
-            return repairKnowledgeGraph(request, progress, cancelRequested);
-        case OnDemandRepairOperation::Fts5:
-            return rebuildFts5Index(request, progress, cancelRequested);
-        case OnDemandRepairOperation::Embeddings:
-            return generateMissingEmbeddings(request, progress, cancelRequested);
-        case OnDemandRepairOperation::Topology:
-            return rebuildTopologyArtifacts(request, progress);
-        case OnDemandRepairOperation::Optimize:
-            return optimizeDatabase(request.dryRun, request.verbose, progress);
+    static_assert(static_cast<std::uint64_t>(OnDemandRepairOperation::Optimize) + 1 ==
+                      metrics::kRepairOperationCodes.size(),
+                  "OnDemandRepairOperation must mirror metrics::kRepairOperationCodes");
+    auto* op = repair::repairOperationForCode(static_cast<std::uint64_t>(operation) + 1);
+    if (!op) {
+        RepairOperationResult result;
+        result.operation = "unknown";
+        result.failed = 1;
+        result.message = "Unknown repair operation";
+        return result;
     }
-    RepairOperationResult result;
-    result.operation = "unknown";
-    result.failed = 1;
-    result.message = "Unknown repair operation";
-    return result;
+    repair::OperationEnv env{ctx_, cfg_, state_, running_};
+    return op->run(env, request, progress, cancelRequested);
 }
 
 RepairResponse RepairService::executeRepair(const RepairRequest& request, ProgressFn progress,
@@ -1243,11 +1049,10 @@ RepairService::executeRepairAsync(const RepairRequest& request, ProgressFn progr
         }
 
         RepairOperationResult result;
-        if (!request.foreground && operation == OnDemandRepairOperation::StuckDocuments) {
-            result = co_await recoverStuckDocumentsAsync(request, run.progress(), cancelRequested);
-        } else if (!request.foreground && operation == OnDemandRepairOperation::Embeddings) {
-            result =
-                co_await generateMissingEmbeddingsAsync(request, run.progress(), cancelRequested);
+        auto* op = repair::repairOperationForCode(static_cast<std::uint64_t>(operation) + 1);
+        if (op && !request.foreground) {
+            repair::OperationEnv env{ctx_, cfg_, state_, running_};
+            result = co_await op->runAsync(env, request, run.progress(), cancelRequested);
         } else {
             result = runRepairOperation(operation, request, run.progress(), cancelRequested);
         }
@@ -1263,2170 +1068,8 @@ RepairService::executeRepairAsync(const RepairRequest& request, ProgressFn progr
 // Stuck Document Recovery (NEW)
 // ============================================================================
 
-std::vector<RepairService::StuckDocumentInfo>
-RepairService::detectStuckDocuments(int32_t maxRetries) {
-    std::vector<StuckDocumentInfo> stuck;
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta)
-        return stuck;
-
-    auto now = std::chrono::system_clock::now();
-    auto threshold = std::chrono::duration_cast<std::chrono::seconds>(cfg_.stalledThreshold);
-    auto cutoffEpoch =
-        std::chrono::duration_cast<std::chrono::seconds>((now - threshold).time_since_epoch())
-            .count();
-
-    // Use targeted SQL queries for each stuck category instead of loading all docs.
-
-    // Category 1: Failed extraction
-    {
-        metadata::DocumentQueryOptions opts;
-        opts.extractionStatuses = {metadata::ExtractionStatus::Failed};
-        opts.maxRepairAttempts = maxRetries;
-
-        auto docsResult = meta->queryDocuments(opts);
-        if (docsResult) {
-            for (const auto& d : docsResult.value()) {
-                stuck.push_back({StuckDocumentInfo::FailedExtraction, d.id, d.sha256Hash,
-                                 d.filePath, d.repairAttempts});
-            }
-        }
-    }
-
-    // Category 2: Ghost success (Success but no content row)
-    {
-        metadata::DocumentQueryOptions opts;
-        opts.extractionStatuses = {metadata::ExtractionStatus::Success};
-        opts.maxRepairAttempts = maxRetries;
-        opts.onlyMissingContent = true;
-
-        auto docsResult = meta->queryDocuments(opts);
-        if (docsResult) {
-            for (const auto& d : docsResult.value()) {
-                stuck.push_back({StuckDocumentInfo::GhostSuccess, d.id, d.sha256Hash, d.filePath,
-                                 d.repairAttempts});
-            }
-        }
-    }
-
-    // Category 3: Stalled Pending (indexed/modified before cutoff)
-    {
-        metadata::DocumentQueryOptions opts;
-        opts.extractionStatuses = {metadata::ExtractionStatus::Pending};
-        opts.maxRepairAttempts = maxRetries;
-        opts.stalledBefore = cutoffEpoch;
-
-        auto docsResult = meta->queryDocuments(opts);
-        if (docsResult) {
-            for (const auto& d : docsResult.value()) {
-                stuck.push_back({StuckDocumentInfo::StalledPending, d.id, d.sha256Hash, d.filePath,
-                                 d.repairAttempts});
-            }
-        }
-    }
-
-    // Category 4: Stalled Processing (repairStatus stuck in Processing)
-    {
-        metadata::DocumentQueryOptions opts;
-        opts.repairStatuses = {metadata::RepairStatus::Processing};
-        opts.maxRepairAttempts = maxRetries;
-        opts.repairAttemptedBefore = cutoffEpoch;
-
-        auto docsResult = meta->queryDocuments(opts);
-        if (docsResult) {
-            for (const auto& d : docsResult.value()) {
-                stuck.push_back({StuckDocumentInfo::StalledProcessing, d.id, d.sha256Hash,
-                                 d.filePath, d.repairAttempts});
-            }
-        }
-    }
-
-    return stuck;
-}
-
-boost::asio::awaitable<RepairOperationResult>
-RepairService::recoverStuckDocumentsAsync(const RepairRequest& req, const ProgressFn& progress,
-                                          std::atomic<bool>* cancelRequested) {
-    RepairOperationResult result;
-    result.operation = "stuck_docs";
-
-    auto isCanceled = [&]() {
-        return cancelRequested && cancelRequested->load(std::memory_order_relaxed);
-    };
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata repository not available";
-        co_return result;
-    }
-
-    auto* wc = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-    MetadataWriteFacade metaFacade(wc, meta.get());
-
-    auto stuckDocs = detectStuckDocuments(req.maxRetries);
-    result.processed = stuckDocs.size();
-
-    if (stuckDocs.empty()) {
-        result.message = "No stuck documents found";
-        co_return result;
-    }
-
-    spdlog::info("RepairService: found {} stuck documents", stuckDocs.size());
-
-    if (req.dryRun) {
-        result.message = "Would recover " + std::to_string(stuckDocs.size()) + " stuck documents";
-        result.skipped = stuckDocs.size();
-        co_return result;
-    }
-
-    const std::size_t rpcCapacity = static_cast<std::size_t>(TuneAdvisor::postIngestRpcQueueMax());
-    const bool useRpcChannel = (rpcCapacity > 0);
-    auto postIngestChannel =
-        InternalEventBus::instance().get_or_create_channel<InternalEventBus::PostIngestTask>(
-            useRpcChannel ? "post_ingest_rpc" : "post_ingest",
-            useRpcChannel ? rpcCapacity : std::size_t(4096));
-    boost::asio::steady_timer timer(co_await boost::asio::this_coro::executor);
-
-    for (std::size_t i = 0; i < stuckDocs.size(); ++i) {
-        if (isCanceled()) {
-            result.failed += (stuckDocs.size() - i);
-            result.message = "Repair canceled";
-            co_return result;
-        }
-
-        const auto& s = stuckDocs[i];
-        InternalEventBus::PostIngestTask task;
-        task.hash = s.hash;
-        task.mime = "";
-
-        const std::string jobLabel =
-            std::string(useRpcChannel ? "post_ingest_rpc" : "post_ingest") + " stuck-doc " +
-            std::to_string(i + 1) + "/" + std::to_string(stuckDocs.size());
-        auto notifyQueueFull = [&progress]() {
-            if (!progress)
-                return;
-            RepairEvent event;
-            event.phase = "repairing";
-            event.operation = "stuck_docs";
-            event.message = "Waiting for post-ingest queue capacity";
-            progress(event);
-        };
-        const bool pushed =
-            co_await queueWithBackoff(postIngestChannel, std::move(task), timer, running_, jobLabel,
-                                      1, 500, 500, isCanceled, notifyQueueFull);
-        if (pushed) {
-            TuningManager::notifyWakeup();
-            auto docRes = meta->getDocument(s.docId);
-            if (docRes && docRes.value().has_value()) {
-                auto doc = docRes.value().value();
-                doc.repairAttempts = s.repairAttempts + 1;
-                doc.repairAttemptedAt = std::chrono::time_point_cast<std::chrono::seconds>(
-                    std::chrono::system_clock::now());
-                metadata::MetadataOpScope opScope("repair_attempt_update");
-                (void)meta->updateDocument(doc);
-            }
-
-            metaFacade.updateExtractionStatus(s.docId, false, metadata::ExtractionStatus::Pending,
-                                              "RepairService: recovery attempt");
-            submitRepairStatusUpdate(ctx_, meta, std::vector<std::string>{s.hash},
-                                     metadata::RepairStatus::Pending,
-                                     "RepairService::recovery/single");
-            ++result.succeeded;
-        } else if (isCanceled()) {
-            result.failed += (stuckDocs.size() - i);
-            result.message = "Repair canceled";
-            co_return result;
-        } else {
-            ++result.failed;
-            spdlog::warn("RepairService: PostIngest channel full after retries, could not "
-                         "re-enqueue {} (consider increasing YAMS_POST_INGEST_RPC_QUEUE_MAX)",
-                         s.hash.substr(0, 12));
-        }
-
-        if (progress && ((i + 1) % 10 == 0 || i + 1 == stuckDocs.size())) {
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "stuck_docs";
-            ev.processed = i + 1;
-            ev.total = stuckDocs.size();
-            ev.succeeded = result.succeeded;
-            ev.failed = result.failed;
-            progress(ev);
-        }
-    }
-    metaFacade.flush();
-
-    result.message = "Recovered " + std::to_string(result.succeeded) + " stuck documents";
-    if (result.failed > 0) {
-        result.message += " (" + std::to_string(result.failed) +
-                          " could not be enqueued - re-run repair to retry)";
-    }
-    co_return result;
-}
-
-RepairOperationResult RepairService::recoverStuckDocuments(const RepairRequest& req,
-                                                           const ProgressFn& progress) {
-    RepairOperationResult result;
-    result.operation = "stuck_docs";
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata repository not available";
-        return result;
-    }
-
-    auto* wc = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-    MetadataWriteFacade metaFacade(wc, meta.get());
-
-    auto stuckDocs = detectStuckDocuments(req.maxRetries);
-    result.processed = stuckDocs.size();
-
-    if (stuckDocs.empty()) {
-        result.message = "No stuck documents found";
-        return result;
-    }
-
-    spdlog::info("RepairService: found {} stuck documents", stuckDocs.size());
-
-    if (req.dryRun) {
-        result.message = "Would recover " + std::to_string(stuckDocs.size()) + " stuck documents";
-        result.skipped = stuckDocs.size();
-        return result;
-    }
-
-    // Re-enqueue stuck documents for re-extraction via a high-priority PostIngestQueue channel.
-    // Use back-pressure retries with drain polling to avoid channel overflow and silent doc loss.
-    const std::size_t rpcCapacity = static_cast<std::size_t>(TuneAdvisor::postIngestRpcQueueMax());
-    const bool useRpcChannel = (rpcCapacity > 0);
-    auto postIngestChannel =
-        InternalEventBus::instance().get_or_create_channel<InternalEventBus::PostIngestTask>(
-            useRpcChannel ? "post_ingest_rpc" : "post_ingest",
-            useRpcChannel ? rpcCapacity : std::size_t(4096));
-
-    constexpr int kMaxRetryRounds = 20; // 20 rounds × 500ms = 10s max wait per document
-    constexpr auto kDrainPollInterval = std::chrono::milliseconds(500);
-
-    for (size_t i = 0; i < stuckDocs.size(); ++i) {
-        const auto& s = stuckDocs[i];
-
-        // Re-enqueue for extraction (do not mutate metadata unless enqueue succeeds).
-        InternalEventBus::PostIngestTask task;
-        task.hash = s.hash;
-        task.mime = ""; // will be re-detected
-
-        // Attempt the push with back-pressure retries
-        bool pushed = postIngestChannel->try_push(task);
-        if (!pushed) {
-            // Channel full — wait for consumer to drain, then retry
-            for (int retry = 0; retry < kMaxRetryRounds && !pushed; ++retry) {
-                std::this_thread::sleep_for(kDrainPollInterval);
-                pushed = postIngestChannel->try_push(task);
-            }
-        }
-
-        if (pushed) {
-            TuningManager::notifyWakeup();
-            // Increment repair attempts
-            auto docRes = meta->getDocument(s.docId);
-            if (docRes && docRes.value().has_value()) {
-                auto doc = docRes.value().value();
-                doc.repairAttempts = s.repairAttempts + 1;
-                doc.repairAttemptedAt = std::chrono::time_point_cast<std::chrono::seconds>(
-                    std::chrono::system_clock::now());
-                metadata::MetadataOpScope opScope("repair_attempt_update");
-                (void)meta->updateDocument(doc);
-            }
-
-            // Reset extraction status to Pending
-            metaFacade.updateExtractionStatus(s.docId, false, metadata::ExtractionStatus::Pending,
-                                              "RepairService: recovery attempt");
-
-            submitRepairStatusUpdate(ctx_, meta, std::vector<std::string>{s.hash},
-                                     metadata::RepairStatus::Pending,
-                                     "RepairService::repairAttempt/reset");
-
-            ++result.succeeded;
-        } else {
-            ++result.failed;
-            spdlog::warn("RepairService: PostIngest channel full after retries, could not "
-                         "re-enqueue {} (consider increasing YAMS_POST_INGEST_RPC_QUEUE_MAX)",
-                         s.hash.substr(0, 12));
-        }
-
-        if (progress && (i + 1) % 10 == 0) {
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "stuck_docs";
-            ev.processed = i + 1;
-            ev.total = stuckDocs.size();
-            ev.succeeded = result.succeeded;
-            ev.failed = result.failed;
-            progress(ev);
-        }
-    }
-
-    result.message = "Recovered " + std::to_string(result.succeeded) + " stuck documents";
-    if (result.failed > 0) {
-        result.message += " (" + std::to_string(result.failed) +
-                          " could not be enqueued — re-run repair to retry)";
-    }
-    metaFacade.flush();
-    return result;
-}
-
 // ============================================================================
 // Core Repair Operations
 // ============================================================================
-
-RepairOperationResult RepairService::cleanOrphanedMetadata(bool dryRun, bool verbose,
-                                                           bool removeCorrupt,
-                                                           ProgressFn progress) {
-    RepairOperationResult result;
-    result.operation = "orphans";
-
-    auto store = ctx_.getContentStore ? ctx_.getContentStore() : nullptr;
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!store || !meta) {
-        result.message = "Store or metadata not available";
-        return result;
-    }
-
-    // Phase 1: Quick-check if salvage is needed, then optionally salvage documents
-    // from corrupt metadata DBs before scanning for orphans.  Skip salvage entirely
-    // when all corrupt-DB documents are already present in the current DB.
-    size_t salvagedCount = 0;
-    bool salvageDidRun = false;
-    if (!dryRun) {
-        namespace fs = std::filesystem;
-        fs::path dbPath = cfg_.dataDir / "yams.db";
-        if (fs::exists(dbPath)) {
-            auto qc = quickCheckSalvageNeeded(cfg_.dataDir, dbPath);
-            if (qc.needsSalvage) {
-                spdlog::info("[RepairService] Salvage needed: corrupt DB(s) have {} doc(s) "
-                             "more than current DB ({})",
-                             qc.maxCorruptCount - qc.currentDocCount, qc.currentDocCount);
-                SalvageProgressFn salvageProgress;
-                if (progress) {
-                    salvageProgress = [&progress](const std::string& phase,
-                                                  const std::string& message, uint64_t proc,
-                                                  uint64_t total) {
-                        RepairEvent ev;
-                        ev.phase = phase;
-                        ev.operation = "orphans";
-                        ev.processed = proc;
-                        ev.total = total;
-                        ev.message = message;
-                        progress(ev);
-                    };
-                }
-                auto salvaged = salvageFromAllCorruptDbs(cfg_.dataDir, dbPath, salvageProgress);
-                salvagedCount = salvaged.combined.documentsSalvaged;
-                salvageDidRun = true;
-                if (salvagedCount > 0) {
-                    spdlog::info("[RepairService] Salvage recovered {} document(s) "
-                                 "from {} corrupt DB(s)",
-                                 salvagedCount, salvaged.salvagedPaths.size());
-                }
-            } else {
-                spdlog::info("[RepairService] Salvage not needed "
-                             "(all {} docs already in current DB)",
-                             qc.currentDocCount);
-            }
-        }
-    }
-
-    // Phase 2: Scan for orphaned metadata entries (docs without CAS blocks)
-    auto docsResult = metadata::queryDocumentsByPattern(*meta, "%");
-    if (!docsResult) {
-        result.message = "Failed to query: " + docsResult.error().message;
-        return result;
-    }
-
-    std::vector<int64_t> orphanedIds;
-    for (const auto& doc : docsResult.value()) {
-        auto existsResult = store->exists(doc.sha256Hash);
-        if (!existsResult || !existsResult.value())
-            orphanedIds.push_back(doc.id);
-    }
-
-    result.processed = docsResult.value().size();
-
-    if (orphanedIds.empty()) {
-        result.message = "No orphaned entries";
-    } else if (dryRun) {
-        result.skipped = orphanedIds.size();
-        result.message = "Would clean " + std::to_string(orphanedIds.size()) + " orphans";
-    } else {
-        metadata::MetadataOpScope opScope("repair_orphan_cleanup");
-        auto batchResult = meta->deleteDocumentsBatch(orphanedIds);
-        if (batchResult) {
-            result.succeeded = batchResult.value();
-        } else {
-            for (int64_t id : orphanedIds) {
-                auto del = meta->deleteDocument(id);
-                if (del)
-                    result.succeeded++;
-                else
-                    result.failed++;
-            }
-        }
-        result.message = "Cleaned " + std::to_string(result.succeeded) + " orphans";
-    }
-
-    if (salvagedCount > 0) {
-        result.message +=
-            ", salvaged " + std::to_string(salvagedCount) + " documents from corrupt DBs";
-    }
-
-    // Phase 3: If requested, remove the corrupt DB files now that salvage is complete
-    if (!dryRun && removeCorrupt && salvageDidRun) {
-        auto cleanup = removeCorruptDbFiles(cfg_.dataDir);
-        if (!cleanup.removed.empty()) {
-            spdlog::info("[RepairService] Removed {} corrupt DB file(s)", cleanup.removed.size());
-            result.message +=
-                ", removed " + std::to_string(cleanup.removed.size()) + " corrupt DB file(s)";
-        }
-        for (const auto& err : cleanup.errors) {
-            spdlog::warn("[RepairService] Corrupt DB removal error: {}", err);
-        }
-    }
-
-    return result;
-}
-
-RepairOperationResult RepairService::repairMimeTypes(bool dryRun, bool verbose,
-                                                     ProgressFn progress) {
-    RepairOperationResult result;
-    result.operation = "mime";
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        return result;
-    }
-
-    auto* wc = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-    MetadataWriteFacade metaFacade(wc, meta.get());
-
-    auto store = ctx_.getContentStore ? ctx_.getContentStore() : nullptr;
-    (void)detection::FileTypeDetector::initializeWithMagicNumbers();
-    auto& detector = detection::FileTypeDetector::instance();
-
-    constexpr int kBatchSize = 5000;
-    constexpr uint64_t kProgressStride = 100;
-    const auto emit = [&](std::string phase, uint64_t processed, const std::string& message) {
-        if (!progress)
-            return;
-        RepairEvent ev;
-        ev.phase = std::move(phase);
-        ev.operation = "mime";
-        ev.processed = processed;
-        ev.succeeded = result.succeeded;
-        ev.failed = result.failed;
-        ev.skipped = result.skipped;
-        ev.message = message;
-        progress(ev);
-    };
-
-    uint64_t totalScanned = 0;
-    uint64_t totalCandidates = 0;
-    int offset = 0;
-    while (true) {
-        metadata::DocumentQueryOptions opts;
-        opts.limit = kBatchSize;
-        opts.offset = offset;
-        auto batchResult = meta->queryDocuments(opts);
-        if (!batchResult) {
-            result.message = "Failed to query: " + batchResult.error().message;
-            return result;
-        }
-        const auto& batch = batchResult.value();
-        if (batch.empty())
-            break;
-
-        std::vector<std::pair<int64_t, std::string>> toRepair;
-        toRepair.reserve(batch.size());
-        for (const auto& doc : batch) {
-            ++totalScanned;
-            if (shouldRedetectMime(doc)) {
-                const auto detectedMime = bestEffortMimeForDocument(doc, store);
-                if (!detectedMime.empty() && detectedMime != doc.mimeType) {
-                    toRepair.push_back({doc.id, detectedMime});
-                    ++totalCandidates;
-                }
-            }
-            if (totalScanned % kProgressStride == 0) {
-                emit("repairing", totalScanned, "Scanning documents");
-            }
-        }
-
-        if (dryRun) {
-            result.skipped += toRepair.size();
-        } else {
-            for (auto& [id, mimeType] : toRepair) {
-                auto docResult = meta->getDocument(id);
-                if (!(docResult && docResult.value())) {
-                    result.failed++;
-                    continue;
-                }
-
-                auto doc = *docResult.value();
-                const bool oldWasText = detector.isTextMimeType(doc.mimeType);
-                const bool newIsText = detector.isTextMimeType(mimeType);
-                doc.mimeType = std::move(mimeType);
-
-                metadata::MetadataOpScope opScope("repair_mime_update");
-                if (meta->updateDocument(doc)) {
-                    if (oldWasText && !newIsText) {
-                        (void)meta->deleteContent(id);
-                        (void)meta->removeFromIndex(id);
-                        metaFacade.updateExtractionStatus(
-                            id, false, metadata::ExtractionStatus::Pending,
-                            "MIME repaired; stale text content cleared");
-                    }
-                    result.succeeded++;
-                } else {
-                    result.failed++;
-                }
-                if ((result.succeeded + result.failed) % kProgressStride == 0) {
-                    emit("repairing", totalScanned, "Repairing MIME types");
-                }
-            }
-        }
-
-        emit("repairing", totalScanned, "Processed batch");
-
-        if (static_cast<int>(batch.size()) < kBatchSize)
-            break;
-        offset += kBatchSize;
-    }
-
-    result.processed = totalScanned;
-
-    if (totalCandidates == 0) {
-        result.message = "All documents have valid MIME types";
-    } else if (dryRun) {
-        result.message = "Would repair " + std::to_string(result.skipped) + " MIME types";
-    } else {
-        result.message = "Repaired " + std::to_string(result.succeeded) + " MIME types";
-    }
-    metaFacade.flush();
-    return result;
-}
-
-RepairOperationResult RepairService::repairDownloads(bool dryRun, bool verbose,
-                                                     ProgressFn progress) {
-    RepairOperationResult result;
-    result.operation = "downloads";
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        return result;
-    }
-
-    auto* wc = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-    MetadataWriteFacade metaFacade(wc, meta.get());
-
-    auto docsResult = metadata::queryDocumentsByPattern(*meta, "%");
-    if (!docsResult) {
-        result.message = "Failed to query";
-        return result;
-    }
-
-    auto is_url = [](const std::string& s) { return s.find("://") != std::string::npos; };
-    auto extract_host = [](const std::string& url) -> std::string {
-        auto p = url.find("://");
-        if (p == std::string::npos)
-            return {};
-        auto rest = url.substr(p + 3);
-        auto slash = rest.find('/');
-        return (slash == std::string::npos) ? rest : rest.substr(0, slash);
-    };
-    auto extract_scheme = [](const std::string& url) -> std::string {
-        auto p = url.find("://");
-        return (p == std::string::npos) ? std::string{} : url.substr(0, p);
-    };
-    auto filename_from_url = [](std::string url) -> std::string {
-        auto lastSlash = url.find_last_of('/');
-        if (lastSlash != std::string::npos)
-            url = url.substr(lastSlash + 1);
-        auto q = url.find('?');
-        if (q != std::string::npos)
-            url = url.substr(0, q);
-        if (url.empty())
-            url = "downloaded_file";
-        return url;
-    };
-
-    for (auto& doc : docsResult.value()) {
-        std::string sourceUrl;
-        if (is_url(doc.filePath))
-            sourceUrl = doc.filePath;
-        if (sourceUrl.empty())
-            continue;
-
-        result.processed++;
-        if (dryRun) {
-            result.skipped++;
-            continue;
-        }
-
-        std::string filename = filename_from_url(sourceUrl);
-        std::string ext;
-        auto dotPos = filename.rfind('.');
-        if (dotPos != std::string::npos)
-            ext = filename.substr(dotPos);
-        doc.fileName = filename;
-        doc.filePath = std::move(filename);
-        doc.fileExtension = std::move(ext);
-
-        {
-            metadata::MetadataOpScope opScope("repair_download_url_update");
-            if (auto up = meta->updateDocument(doc); up) {
-                result.succeeded++;
-            } else {
-                result.failed++;
-            }
-        }
-
-        try {
-            metaFacade.setMetadata(doc.id, "source_url", metadata::MetadataValue(sourceUrl));
-            metaFacade.setMetadata(doc.id, "tag:downloaded", metadata::MetadataValue("downloaded"));
-            auto host = extract_host(sourceUrl);
-            auto scheme = extract_scheme(sourceUrl);
-            if (!host.empty())
-                metaFacade.setMetadata(doc.id, "tag:host:" + host,
-                                       metadata::MetadataValue("host:" + host));
-            if (!scheme.empty())
-                metaFacade.setMetadata(doc.id, "tag:scheme:" + scheme,
-                                       metadata::MetadataValue("scheme:" + scheme));
-        } catch (const std::exception& e) {
-            spdlog::debug("RepairService: failed to persist download metadata for {}: {}",
-                          sourceUrl, e.what());
-        } catch (...) {
-            spdlog::debug("RepairService: failed to persist download metadata for {}", sourceUrl);
-        }
-    }
-
-    result.message = "Updated " + std::to_string(result.succeeded) + " download documents";
-    metaFacade.flush();
-    return result;
-}
-
-RepairOperationResult RepairService::applySemanticDedupe(const RepairRequest& req,
-                                                         const ProgressFn& progress) {
-    RepairOperationResult result;
-    result.operation = "dedupe";
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        return result;
-    }
-
-    auto groupsResult = meta->listSemanticDuplicateGroups(1000);
-    if (!groupsResult) {
-        result.message =
-            "Failed to load semantic duplicate groups: " + groupsResult.error().message;
-        return result;
-    }
-
-    std::vector<metadata::SemanticDuplicateGroup> groups;
-    groups.reserve(groupsResult.value().size());
-    for (const auto& group : groupsResult.value()) {
-        if (group.status == "suggested") {
-            groups.push_back(group);
-        }
-    }
-
-    result.processed = groups.size();
-    if (groups.empty()) {
-        result.message = "No semantic duplicate suggestions to apply";
-        return result;
-    }
-
-    std::vector<int64_t> canonicalIds;
-    canonicalIds.reserve(groups.size());
-    for (const auto& group : groups) {
-        if (group.canonicalDocumentId.has_value()) {
-            canonicalIds.push_back(*group.canonicalDocumentId);
-        }
-    }
-
-    auto detailsResult = meta->getSemanticDuplicateGroupsForDocuments(canonicalIds);
-    if (!detailsResult) {
-        result.message =
-            "Failed to load semantic duplicate members: " + detailsResult.error().message;
-        return result;
-    }
-
-    for (size_t i = 0; i < groups.size(); ++i) {
-        const auto& group = groups[i];
-        if (!group.canonicalDocumentId.has_value()) {
-            result.failed++;
-            continue;
-        }
-
-        auto detailIt = detailsResult.value().find(*group.canonicalDocumentId);
-        if (detailIt == detailsResult.value().end()) {
-            result.failed++;
-            continue;
-        }
-
-        std::vector<int64_t> toDelete;
-        for (const auto& member : detailIt->second.members) {
-            if (member.role != "canonical") {
-                toDelete.push_back(member.documentId);
-            }
-        }
-
-        if (req.dryRun) {
-            result.skipped += toDelete.size();
-        } else {
-            metadata::MetadataOpScope opScope("repair_dedupe_delete");
-            if (!toDelete.empty()) {
-                auto batchDelete = meta->deleteDocumentsBatch(toDelete);
-                if (batchDelete) {
-                    result.succeeded += batchDelete.value();
-                    if (batchDelete.value() < toDelete.size()) {
-                        result.failed += (toDelete.size() - batchDelete.value());
-                    }
-                } else {
-                    for (int64_t docId : toDelete) {
-                        auto del = meta->deleteDocument(docId);
-                        if (del)
-                            result.succeeded++;
-                        else
-                            result.failed++;
-                    }
-                }
-            }
-
-            auto statusResult = meta->updateSemanticDuplicateGroupStatus(group.groupKey, "applied");
-            if (!statusResult) {
-                result.failed++;
-            }
-        }
-
-        if (progress) {
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "dedupe";
-            ev.processed = i + 1;
-            ev.total = groups.size();
-            ev.succeeded = result.succeeded;
-            ev.failed = result.failed;
-            ev.skipped = result.skipped;
-            ev.message = req.dryRun ? "Previewing semantic duplicate removals"
-                                    : "Removing semantic duplicates";
-            progress(ev);
-        }
-    }
-
-    if (req.dryRun) {
-        result.message =
-            "Would remove " + std::to_string(result.skipped) + " semantic duplicate documents";
-    } else {
-        result.message =
-            "Removed " + std::to_string(result.succeeded) + " semantic duplicate documents";
-    }
-    return result;
-}
-
-RepairOperationResult RepairService::rebuildPathTree(bool dryRun, bool verbose,
-                                                     ProgressFn progress) {
-    RepairOperationResult result;
-    result.operation = "path_tree";
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        return result;
-    }
-
-    // Use RepairManager for path tree repair
-    auto concreteRepo = std::dynamic_pointer_cast<metadata::MetadataRepository>(meta);
-    if (!concreteRepo) {
-        result.message = "Could not get concrete MetadataRepository";
-        return result;
-    }
-
-    if (dryRun) {
-        auto countResult = concreteRepo->countDocsMissingPathTree();
-        if (!countResult) {
-            result.message = "Failed to count missing path tree entries";
-            return result;
-        }
-        uint64_t missing = countResult.value();
-        result.processed = missing;
-        result.skipped = missing;
-        result.message = "Would rebuild " + std::to_string(missing) + " path tree entries";
-        return result;
-    }
-
-    integrity::RepairManager repairMgr(*concreteRepo);
-    auto repairResult = repairMgr.repairPathTree([](uint64_t, uint64_t) {});
-    if (!repairResult) {
-        result.message = "Path tree repair failed: " + repairResult.error().message;
-        result.failed = 1;
-        return result;
-    }
-
-    const auto& stats = repairResult.value();
-    result.processed = stats.documentsScanned;
-    result.succeeded = stats.nodesCreated;
-    result.failed = stats.errors;
-    result.message = "Created " + std::to_string(stats.nodesCreated) + " path tree entries";
-    return result;
-}
-
-RepairOperationResult RepairService::repairKnowledgeGraph(const RepairRequest& req,
-                                                          const ProgressFn& progress,
-                                                          std::atomic<bool>* cancelRequested) {
-    RepairOperationResult result;
-    result.operation = "graph";
-
-    auto graphComponent = ctx_.getGraphComponent ? ctx_.getGraphComponent() : nullptr;
-    auto kgStore = ctx_.getKgStore ? ctx_.getKgStore() : nullptr;
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!graphComponent || !kgStore || !meta) {
-        result.message = "Knowledge graph components not available";
-        return result;
-    }
-
-    auto graphProgress = [&progress](uint64_t processed, uint64_t total,
-                                     const GraphComponent::RepairStats& snap) {
-        if (!progress)
-            return;
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "graph";
-        ev.processed = processed;
-        ev.total = total;
-        ev.succeeded = snap.nodesCreated + snap.nodesUpdated + snap.edgesCreated;
-        ev.failed = snap.errors;
-        ev.message = "graph repair scanning";
-        progress(ev);
-    };
-    auto repairRes = graphComponent->repairGraph(req.dryRun, graphProgress, cancelRequested);
-    if (!repairRes) {
-        result.failed = 1;
-        result.message = "Graph repair failed: " + repairRes.error().message;
-        return result;
-    }
-
-    const auto& stats = repairRes.value();
-    result.processed += stats.nodesCreated + stats.nodesUpdated + stats.edgesCreated;
-    result.succeeded += stats.nodesCreated + stats.nodesUpdated + stats.edgesCreated;
-    result.failed += stats.errors;
-    for (const auto& issue : stats.issues) {
-        spdlog::warn("[RepairService] graph repair issue: {}", issue);
-    }
-
-    auto cleanup = cleanOrphanedKgEntries(req.dryRun, req.verbose, progress);
-    result.processed += cleanup.nodesScanned;
-    result.succeeded += cleanup.nodesDeleted + cleanup.edgesDeleted + cleanup.docEntitiesDeleted;
-    result.skipped += cleanup.skipped;
-    result.failed += cleanup.errors;
-    for (const auto& issue : cleanup.issues) {
-        spdlog::warn("[RepairService] graph cleanup issue: {}", issue);
-    }
-
-    std::string message = "Rebuilt graph nodes/edges (nodes=" + std::to_string(stats.nodesCreated) +
-                          ", edges=" + std::to_string(stats.edgesCreated) + ")";
-    if (cleanup.nodesDeleted > 0 || cleanup.edgesDeleted > 0 || cleanup.docEntitiesDeleted > 0) {
-        message += "; cleaned orphans (nodes=" + std::to_string(cleanup.nodesDeleted) +
-                   ", edges=" + std::to_string(cleanup.edgesDeleted) +
-                   ", doc_entities=" + std::to_string(cleanup.docEntitiesDeleted) + ")";
-    }
-    // Surface the graph repair summaries (reconciliation counts, orphaned-node removal, skips)
-    // to the caller — previously these were only logged, so the CLI showed none of them.
-    for (const auto& issue : stats.issues) {
-        message += "; " + issue;
-    }
-    if (req.dryRun) {
-        message += " (dry-run: changes were not committed)";
-    }
-    result.message = std::move(message);
-    return result;
-}
-
-RepairOperationResult RepairService::rebuildTopologyArtifacts(const RepairRequest& req,
-                                                              const ProgressFn& progress) {
-    (void)progress;
-
-    RepairOperationResult result;
-    result.operation = "topology";
-
-    if (!ctx_.rebuildTopologyArtifacts) {
-        result.failed = 1;
-        result.message = "Topology rebuild callback unavailable";
-        return result;
-    }
-
-    auto rebuildResult = ctx_.rebuildTopologyArtifacts(
-        req.dryRun ? std::string{"repair.topology.dry_run"} : std::string{"repair.topology"},
-        req.dryRun, {});
-    if (!rebuildResult) {
-        result.failed = 1;
-        result.message = "Topology rebuild failed: " + rebuildResult.error().message;
-        return result;
-    }
-
-    const auto& stats = rebuildResult.value();
-    result.processed = stats.documentsProcessed;
-    if (stats.skipped || req.dryRun) {
-        result.skipped = stats.documentsProcessed;
-    } else {
-        result.succeeded = stats.documentsProcessed;
-    }
-
-    std::string message = (stats.skipped || req.dryRun) ? "Topology rebuild analyzed "
-                                                        : "Topology rebuild stored artifacts for ";
-    message += std::to_string(stats.documentsProcessed) +
-               " docs (clusters=" + std::to_string(stats.clustersBuilt) +
-               ", memberships=" + std::to_string(stats.membershipsBuilt) + ")";
-    if (!stats.snapshotId.empty()) {
-        message += ", snapshot=" + stats.snapshotId;
-    }
-    if (stats.documentsMissingEmbeddings > 0 || stats.documentsMissingGraphNodes > 0) {
-        message += " [missing_embeddings=" + std::to_string(stats.documentsMissingEmbeddings) +
-                   ", missing_graph_nodes=" + std::to_string(stats.documentsMissingGraphNodes) +
-                   "]";
-    }
-    if (!stats.issues.empty()) {
-        message += ": " + stats.issues.front();
-    }
-    result.message = std::move(message);
-    return result;
-}
-
-RepairService::KgCleanupStats RepairService::cleanOrphanedKgEntries(bool dryRun, bool verbose,
-                                                                    const ProgressFn& progress) {
-    (void)verbose;
-    KgCleanupStats stats;
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    auto kgStore = ctx_.getKgStore ? ctx_.getKgStore() : nullptr;
-    if (!meta || !kgStore) {
-        stats.errors = 1;
-        stats.issues.push_back("Metadata or KG store unavailable");
-        return stats;
-    }
-
-    auto emitProgress = [&](const std::string& message) {
-        if (!progress)
-            return;
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "graph";
-        ev.processed = stats.nodesScanned;
-        ev.succeeded = stats.nodesDeleted + stats.edgesDeleted + stats.docEntitiesDeleted;
-        ev.failed = stats.errors;
-        ev.skipped = stats.skipped;
-        ev.message = message;
-        progress(ev);
-    };
-
-    auto scanType = [&](const std::string& type, const std::string& prefix, bool deleteByHash) {
-        constexpr std::size_t kBatchSize = 500;
-        std::size_t offset = 0;
-        std::unique_ptr<WriteBatch> orphanDeleteBatch;
-        {
-            auto* coord = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-            if (coord) {
-                orphanDeleteBatch = std::make_unique<WriteBatch>();
-                orphanDeleteBatch->source = "RepairService::orphanKgNodeDelete";
-            }
-        }
-        while (true) {
-            auto nodesRes = kgStore->findNodesByType(type, kBatchSize, offset);
-            if (!nodesRes) {
-                ++stats.errors;
-                stats.issues.push_back("findNodesByType(" + type +
-                                       ") failed: " + nodesRes.error().message);
-                break;
-            }
-
-            const auto& nodes = nodesRes.value();
-            if (nodes.empty())
-                break;
-
-            std::size_t deletedNodesInBatch = 0;
-            for (const auto& node : nodes) {
-                ++stats.nodesScanned;
-                if (node.nodeKey.compare(0, prefix.size(), prefix) != 0)
-                    continue;
-
-                std::string hash = node.nodeKey.substr(prefix.size());
-                if (hash.empty())
-                    continue;
-
-                auto docRes = meta->getDocumentByHash(hash);
-                if (!docRes) {
-                    ++stats.errors;
-                    stats.issues.push_back("getDocumentByHash failed for " + hash + ": " +
-                                           docRes.error().message);
-                    continue;
-                }
-                if (!docRes.value().has_value()) {
-                    ++stats.orphanNodes;
-                    if (dryRun) {
-                        ++stats.skipped;
-                        continue;
-                    }
-                    if (orphanDeleteBatch) {
-                        if (deleteByHash) {
-                            orphanDeleteBatch->ops.emplace_back(DeleteNodesForDocumentHashOp{hash});
-                        } else {
-                            orphanDeleteBatch->ops.emplace_back(DeleteNodeByIdOp{node.id});
-                        }
-                    } else {
-                        auto* coord =
-                            ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-                        if (!coord) {
-                            ++stats.errors;
-                            stats.issues.push_back(
-                                "WriteCoordinator unavailable for orphan KG node delete");
-                            continue;
-                        }
-                        auto wb = std::make_unique<WriteBatch>();
-                        wb->source = "RepairService::orphanKgNodeDelete";
-                        if (deleteByHash) {
-                            wb->ops.emplace_back(DeleteNodesForDocumentHashOp{hash});
-                        } else {
-                            wb->ops.emplace_back(DeleteNodeByIdOp{node.id});
-                        }
-                        coord->enqueue(std::move(wb));
-                        auto flushRes = coord->flush();
-                        if (!flushRes) {
-                            ++stats.errors;
-                            stats.issues.push_back("orphan KG node delete flush failed for " +
-                                                   node.nodeKey + ": " + flushRes.error().message);
-                            continue;
-                        }
-                    }
-                    ++stats.nodesDeleted;
-                    ++deletedNodesInBatch;
-                }
-
-                if (stats.nodesScanned % 200 == 0) {
-                    emitProgress("Scanned " + std::to_string(stats.nodesScanned) +
-                                 " KG nodes (orphans=" + std::to_string(stats.orphanNodes) + ")");
-                }
-            }
-
-            if (nodes.size() < kBatchSize)
-                break;
-
-            if (!dryRun && deletedNodesInBatch > 0) {
-                if (offset >= deletedNodesInBatch) {
-                    offset -= deletedNodesInBatch;
-                } else {
-                    offset = 0;
-                }
-            }
-            offset += nodes.size();
-        }
-
-        // Flush accumulated orphan node deletes as a single batch.
-        if (orphanDeleteBatch && !orphanDeleteBatch->ops.empty()) {
-            auto* coord = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-            if (coord) {
-                coord->enqueue(std::move(orphanDeleteBatch));
-                auto flushRes = coord->flush();
-                if (!flushRes) {
-                    ++stats.errors;
-                    stats.issues.push_back("orphan KG node delete batch flush failed: " +
-                                           flushRes.error().message);
-                }
-            }
-        }
-    };
-
-    scanType("document", "doc:", true);
-    scanType("blob", "blob:", false);
-
-    if (!dryRun) {
-        auto* coord = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-        if (!coord) {
-            ++stats.errors;
-            stats.issues.push_back(
-                "WriteCoordinator unavailable for orphan edge/doc_entities cleanup");
-        } else {
-            const auto before = coord->getStats();
-            auto wb = std::make_unique<WriteBatch>();
-            wb->source = "RepairService::kgOrphanCleanup";
-            wb->ops.emplace_back(DeleteOrphanedEdgesOp{});
-            wb->ops.emplace_back(DeleteOrphanedDocEntitiesOp{});
-            coord->enqueue(std::move(wb));
-            auto flushRes = coord->flush();
-            if (!flushRes) {
-                ++stats.errors;
-                stats.issues.push_back("orphan edge/doc_entities cleanup flush failed: " +
-                                       flushRes.error().message);
-            } else {
-                // flush() drains the queue, so the coordinator stats now reflect this batch.
-                const auto after = coord->getStats();
-                if (after.edgesDeleted >= before.edgesDeleted)
-                    stats.edgesDeleted += after.edgesDeleted - before.edgesDeleted;
-                if (after.docEntitiesDeleted >= before.docEntitiesDeleted)
-                    stats.docEntitiesDeleted +=
-                        after.docEntitiesDeleted - before.docEntitiesDeleted;
-            }
-        }
-    } else {
-        emitProgress("Dry-run: skipped orphan edge/doc_entities cleanup");
-    }
-
-    if (stats.edgesDeleted > 0 || stats.docEntitiesDeleted > 0) {
-        emitProgress("Cleaned orphan edges=" + std::to_string(stats.edgesDeleted) +
-                     ", stale doc_entities=" + std::to_string(stats.docEntitiesDeleted));
-    }
-
-    return stats;
-}
-
-RepairOperationResult RepairService::cleanOrphanedChunks(bool dryRun, bool verbose,
-                                                         const ProgressFn& progress) {
-    RepairOperationResult result;
-    result.operation = "chunks";
-
-    namespace fs = std::filesystem;
-    fs::path refsDbPath = cfg_.dataDir / "storage" / "refs.db";
-    fs::path objectsPath = cfg_.dataDir / "storage" / "objects";
-
-    if (!fs::exists(refsDbPath) || !fs::exists(objectsPath)) {
-        result.message = "No refs.db or objects directory found";
-        return result;
-    }
-
-    sqlite3* db = nullptr;
-    if (sqlite3_open(refsDbPath.string().c_str(), &db) != SQLITE_OK) {
-        result.message = "Failed to open refs.db";
-        return result;
-    }
-
-    // Quick-check: count unreferenced block_references entries.
-    // If zero, the common case (no orphans expected) can skip the expensive FS walk.
-    {
-        sqlite3_stmt* countStmt = nullptr;
-        const char* countSql = "SELECT COUNT(*) FROM block_references WHERE ref_count = 0";
-        int zeroRefCount = 0;
-        if (sqlite3_prepare_v2(db, countSql, -1, &countStmt, nullptr) == SQLITE_OK) {
-            if (sqlite3_step(countStmt) == SQLITE_ROW)
-                zeroRefCount = sqlite3_column_int(countStmt, 0);
-            sqlite3_finalize(countStmt);
-        }
-
-        if (zeroRefCount == 0) {
-            // Also do a quick FS sanity check: count prefix dirs. If the count matches
-            // the number of referenced hashes' unique prefixes, no orphans are possible.
-            // For now, if refs.db tracks no zero-ref entries, trust it and skip the walk.
-            sqlite3_close(db);
-            result.message = "No orphaned chunks (quick-check: 0 unreferenced blocks)";
-            return result;
-        }
-
-        spdlog::info("[RepairService] chunks: {} unreferenced block_references, proceeding "
-                     "with filesystem scan",
-                     zeroRefCount);
-    }
-
-    // Get referenced hashes
-    std::set<std::string> referencedHashes;
-    sqlite3_stmt* stmt = nullptr;
-    const char* sql = "SELECT block_hash FROM block_references WHERE ref_count > 0";
-    if (sqlite3_prepare_v2(db, sql, -1, &stmt, nullptr) == SQLITE_OK) {
-        while (sqlite3_step(stmt) == SQLITE_ROW) {
-            const char* hash = reinterpret_cast<const char*>(sqlite3_column_text(stmt, 0));
-            if (hash)
-                referencedHashes.insert(hash);
-        }
-        sqlite3_finalize(stmt);
-    }
-
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "chunks";
-        ev.message = "Loaded " + std::to_string(referencedHashes.size()) +
-                     " referenced hashes, scanning filesystem...";
-        progress(ev);
-    }
-
-    // Scan manifests for additional references
-    size_t manifestDirsScanned = 0;
-    if (fs::exists(objectsPath)) {
-        for (const auto& dirEntry : fs::directory_iterator(objectsPath)) {
-            if (!fs::is_directory(dirEntry))
-                continue;
-            ++manifestDirsScanned;
-            if (progress && manifestDirsScanned % 64 == 0) {
-                RepairEvent ev;
-                ev.phase = "repairing";
-                ev.operation = "chunks";
-                ev.message =
-                    "Scanning manifests (" + std::to_string(manifestDirsScanned) + " dirs)...";
-                progress(ev);
-            }
-            for (const auto& fileEntry : fs::directory_iterator(dirEntry.path())) {
-                if (!fs::is_regular_file(fileEntry))
-                    continue;
-                std::string filename = fileEntry.path().filename().string();
-                if (filename.size() > 9 && filename.substr(filename.size() - 9) == ".manifest") {
-                    try {
-                        std::ifstream mf(fileEntry.path(), std::ios::binary);
-                        if (!mf)
-                            continue;
-                        std::vector<char> data((std::istreambuf_iterator<char>(mf)),
-                                               std::istreambuf_iterator<char>());
-                        if (data.size() >= 28 && data[0] == 'Y' && data[1] == 'M' &&
-                            data[2] == 'N' && data[3] == 'F') {
-                            size_t pos = 24;
-                            uint32_t numChunks = 0;
-                            if (pos + 4 <= data.size()) {
-                                std::memcpy(&numChunks, data.data() + pos, 4);
-                                pos += 4;
-                            }
-                            for (uint32_t i = 0; i < numChunks && pos + 4 <= data.size(); ++i) {
-                                uint32_t hashLen = 0;
-                                std::memcpy(&hashLen, data.data() + pos, 4);
-                                pos += 4;
-                                if (hashLen > 0 && hashLen < 256 && pos + hashLen <= data.size()) {
-                                    referencedHashes.insert(
-                                        std::string(data.data() + pos, hashLen));
-                                    pos += hashLen + 16;
-                                } else {
-                                    break;
-                                }
-                            }
-                        }
-                    } catch (const std::exception& e) {
-                        spdlog::debug("RepairService: failed to parse block references for {}: {}",
-                                      fileEntry.path().string(), e.what());
-                    } catch (...) {
-                        spdlog::debug("RepairService: failed to parse block references for {}",
-                                      fileEntry.path().string());
-                    }
-                }
-            }
-        }
-    }
-
-    // Find orphaned chunks
-    std::vector<std::pair<std::string, fs::path>> orphanedChunks;
-    size_t dirsScanned = 0;
-    if (fs::exists(objectsPath)) {
-        for (const auto& dirEntry : fs::directory_iterator(objectsPath)) {
-            if (!fs::is_directory(dirEntry))
-                continue;
-            ++dirsScanned;
-            // Emit progress every 64 prefix dirs to keep the stream alive
-            if (progress && dirsScanned % 64 == 0) {
-                RepairEvent ev;
-                ev.phase = "repairing";
-                ev.operation = "chunks";
-                ev.message = "Scanning directory " + std::to_string(dirsScanned) + "...";
-                progress(ev);
-            }
-            std::string dirName = dirEntry.path().filename().string();
-            for (const auto& fileEntry : fs::directory_iterator(dirEntry.path())) {
-                if (!fs::is_regular_file(fileEntry))
-                    continue;
-                std::string fn = fileEntry.path().filename().string();
-                if (fn.size() > 9 && fn.substr(fn.size() - 9) == ".manifest")
-                    continue;
-                std::string fullHash = dirName + fn;
-                if (referencedHashes.find(fullHash) == referencedHashes.end()) {
-                    orphanedChunks.push_back({fullHash, fileEntry.path()});
-                }
-            }
-        }
-    }
-
-    result.processed = orphanedChunks.size();
-
-    if (orphanedChunks.empty()) {
-        result.message = "No orphaned chunks";
-        sqlite3_close(db);
-        return result;
-    }
-
-    if (dryRun) {
-        result.skipped = orphanedChunks.size();
-        result.message = "Would delete " + std::to_string(orphanedChunks.size()) + " chunks";
-        sqlite3_close(db);
-        return result;
-    }
-
-    for (const auto& [hash, path] : orphanedChunks) {
-        try {
-            fs::remove(path);
-            result.succeeded++;
-        } catch (...) {
-            result.failed++;
-        }
-    }
-
-    // Clean zero-ref entries from refs.db
-    sqlite3_exec(db, "DELETE FROM block_references WHERE ref_count = 0", nullptr, nullptr, nullptr);
-
-    sqlite3_close(db);
-    result.message = "Deleted " + std::to_string(result.succeeded) + " chunks";
-    return result;
-}
-
-RepairOperationResult RepairService::repairBlockReferences(bool dryRun, bool verbose,
-                                                           const ProgressFn& progress) {
-    RepairOperationResult result;
-    result.operation = "block_refs";
-
-    namespace fs = std::filesystem;
-    fs::path objectsPath = cfg_.dataDir / "storage" / "objects";
-    fs::path refsDbPath = cfg_.dataDir / "storage" / "refs.db";
-
-    if (!fs::exists(objectsPath) || !fs::exists(refsDbPath)) {
-        result.message = "No objects directory or refs.db";
-        return result;
-    }
-
-    auto repairResult = integrity::RepairManager::repairBlockReferences(
-        objectsPath, refsDbPath, dryRun, [&progress](uint64_t processed, uint64_t total) {
-            if (!progress)
-                return;
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "block_refs";
-            ev.processed = processed;
-            ev.total = total;
-            ev.message = "Scanning block references";
-            progress(ev);
-        });
-    if (!repairResult) {
-        result.message = "Block refs repair failed: " + repairResult.error().message;
-        result.failed = 1;
-        return result;
-    }
-
-    const auto& stats = repairResult.value();
-    result.processed = stats.blocksScanned;
-    result.succeeded = stats.blocksUpdated;
-    result.skipped = stats.blocksSkipped;
-    result.failed = stats.errors;
-    result.message = "Scanned " + std::to_string(stats.blocksScanned) + ", updated " +
-                     std::to_string(stats.blocksUpdated);
-    return result;
-}
-
-RepairOperationResult RepairService::rebuildFts5Index(const RepairRequest& req,
-                                                      const ProgressFn& progress,
-                                                      std::atomic<bool>* cancelRequested) {
-    RepairOperationResult result;
-    result.operation = "fts5";
-
-    auto isCanceled = [&]() {
-        return cancelRequested && cancelRequested->load(std::memory_order_relaxed);
-    };
-
-    const bool dryRun = req.dryRun;
-    const bool force = req.force;
-
-    auto store = ctx_.getContentStore ? ctx_.getContentStore() : nullptr;
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!store || !meta) {
-        result.message = "Store or metadata not available";
-        return result;
-    }
-
-    auto* wc = ctx_.getWriteCoordinator ? ctx_.getWriteCoordinator() : nullptr;
-    MetadataWriteFacade metaFacade(wc, meta.get());
-
-    auto customExtractors = ctx_.getContentExtractors
-                                ? ctx_.getContentExtractors()
-                                : std::vector<std::shared_ptr<extraction::IContentExtractor>>{};
-
-    // Pre-load all FTS5 rowids so the incremental skip check below is O(1)
-    // per document instead of one SQL round-trip each.
-    std::unordered_set<int64_t> ftsRowIds;
-    if (!force) {
-        auto ftsResult = meta->getFts5IndexedRowIdSet();
-        if (ftsResult)
-            ftsRowIds = std::move(ftsResult.value());
-    }
-
-    const RepairBudget budget{std::max<std::size_t>(1, cfg_.maxBatch),
-                              std::chrono::milliseconds(250)};
-    std::int64_t cursorDocumentId = 0;
-    std::size_t docsSeen = 0;
-    bool finished = false;
-
-    // Cursor pagination avoids materializing the full corpus and gives the daemon
-    // regular scheduling/progress boundaries during large FTS5 repairs.
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "fts5";
-        ev.processed = 0;
-        ev.total = 0;
-        ev.message =
-            "Rebuilding FTS5 index in slices (batch=" + std::to_string(budget.maxDocuments) + ")";
-        progress(ev);
-    }
-
-    while (!finished) {
-        if (isCanceled()) {
-            result.message =
-                "FTS5 rebuild canceled after scanning " + std::to_string(docsSeen) + " documents";
-            return result;
-        }
-
-        RepairSlice slice;
-        slice.cursorDocumentId = cursorDocumentId;
-
-        while (!slice.exhausted(budget)) {
-            metadata::DocumentQueryOptions opts;
-            opts.idGreaterThan = cursorDocumentId;
-            opts.orderByIdAsc = true;
-            const auto remaining = budget.maxDocuments > slice.processed
-                                       ? budget.maxDocuments - slice.processed
-                                       : std::size_t{1};
-            opts.limit = static_cast<int>(std::min<std::size_t>(remaining, 1024));
-
-            auto docs = meta->queryDocuments(opts);
-            if (!docs) {
-                result.message = "Failed to enumerate after document id " +
-                                 std::to_string(cursorDocumentId) + ": " + docs.error().message;
-                result.failed++;
-                return result;
-            }
-            if (docs.value().empty()) {
-                finished = true;
-                break;
-            }
-
-            for (const auto& d : docs.value()) {
-                if (isCanceled()) {
-                    result.message = "FTS5 rebuild canceled after scanning " +
-                                     std::to_string(docsSeen) + " documents";
-                    return result;
-                }
-
-                cursorDocumentId = std::max(cursorDocumentId, d.id);
-                slice.cursorDocumentId = cursorDocumentId;
-                ++slice.processed;
-                ++docsSeen;
-                ++result.processed;
-
-                // Emit progress frequently so the streaming layer has events to send
-                // (prevents client read timeouts during large rebuilds).
-                if (progress && docsSeen % 25 == 0) {
-                    RepairEvent ev;
-                    ev.phase = "repairing";
-                    ev.operation = "fts5";
-                    ev.processed = docsSeen;
-                    ev.total = 0;
-                    ev.succeeded = result.succeeded;
-                    ev.failed = result.failed;
-                    ev.skipped = result.skipped;
-                    ev.message = "Rebuilding FTS5 index (scanned " + std::to_string(docsSeen) + ")";
-                    progress(ev);
-                }
-
-                auto extractionStatus = d.extractionStatus;
-                bool successHasContent = false;
-                if (extractionStatus == metadata::ExtractionStatus::Success) {
-                    auto contentRes = meta->getContent(d.id);
-                    successHasContent = contentRes && contentRes.value().has_value();
-                    if (!successHasContent) {
-                        metaFacade.updateExtractionStatus(d.id, false,
-                                                          metadata::ExtractionStatus::Pending,
-                                                          "Missing content row; reset by repair");
-                        extractionStatus = metadata::ExtractionStatus::Pending;
-                    }
-                }
-
-                const std::string extension = normalizedRepairExtension(d);
-
-                // Re-detect MIME for unhelpful or clearly wrong types.
-                std::string effectiveMime = d.mimeType;
-                if (shouldRedetectMime(d)) {
-                    effectiveMime = bestEffortMimeForDocument(d, store);
-                    auto updated = d;
-                    updated.mimeType = effectiveMime;
-                    metadata::MetadataOpScope opScope("repair_mime_update");
-                    (void)meta->updateDocument(updated);
-                }
-
-                if (dryRun) {
-                    result.skipped++;
-                    continue;
-                }
-
-                // Incremental mode: skip documents that already have successful extraction
-                // with a valid content row AND a matching FTS5 index entry. Use --force to
-                // unconditionally rebuild everything.
-                if (!force && extractionStatus == metadata::ExtractionStatus::Success &&
-                    successHasContent && ftsRowIds.count(d.id)) {
-                    result.skipped++;
-                    continue;
-                }
-
-                try {
-                    auto extractedOpt = yams::extraction::util::extractDocumentText(
-                        store, d.sha256Hash, effectiveMime, extension, customExtractors);
-                    if (extractedOpt && !extractedOpt->empty()) {
-                        metaFacade.updateExtractionStatus(d.id, false,
-                                                          metadata::ExtractionStatus::Pending,
-                                                          "repair fts5 processing");
-
-                        metadata::DocumentContent content;
-                        content.documentId = d.id;
-                        if (extractedOpt->size() > kMaxTextToPersistInMetadataBytes) {
-                            content.contentText =
-                                extractedOpt->substr(0, kMaxTextToPersistInMetadataBytes);
-                        } else {
-                            content.contentText = *extractedOpt;
-                        }
-                        content.extractionMethod = "repair";
-                        auto contentResult = meta->insertContent(content);
-
-                        // Cap text for FTS5 indexing too — SQLite cannot bind strings > ~2 GB
-                        // and very large documents cause "string or blob too big" errors.
-                        const auto& textForIndex =
-                            (extractedOpt->size() > kMaxTextToPersistInMetadataBytes)
-                                ? content.contentText
-                                : *extractedOpt;
-                        auto ir = meta->indexDocumentContent(d.id, d.fileName, textForIndex,
-                                                             effectiveMime);
-
-                        if (ir && contentResult) {
-                            metaFacade.updateExtractionStatus(
-                                d.id, true, metadata::ExtractionStatus::Success, "");
-                            result.succeeded++;
-                        } else {
-                            static const std::string kUnknownFailure = "unknown";
-                            const std::string& failMsg =
-                                !contentResult ? contentResult.error().message
-                                               : (!ir ? ir.error().message : kUnknownFailure);
-                            metaFacade.updateExtractionStatus(
-                                d.id, false, metadata::ExtractionStatus::Failed, failMsg);
-                            result.failed++;
-                        }
-                    } else {
-                        metaFacade.updateExtractionStatus(d.id, false,
-                                                          metadata::ExtractionStatus::Skipped,
-                                                          "No extractable text");
-                        result.skipped++;
-                    }
-                } catch (const std::exception& e) {
-                    metaFacade.updateExtractionStatus(d.id, false,
-                                                      metadata::ExtractionStatus::Failed, e.what());
-                    result.failed++;
-                }
-
-                if (slice.exhausted(budget))
-                    break;
-            }
-
-            if (docs.value().size() < static_cast<std::size_t>(opts.limit)) {
-                finished = true;
-                break;
-            }
-        }
-
-        if (progress && slice.processed > 0) {
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "fts5";
-            ev.processed = docsSeen;
-            ev.total = 0;
-            ev.succeeded = result.succeeded;
-            ev.failed = result.failed;
-            ev.skipped = result.skipped;
-            ev.message = "Completed FTS5 repair slice through document id " +
-                         std::to_string(slice.cursorDocumentId);
-            progress(ev);
-        }
-        std::this_thread::yield();
-    }
-
-    result.message = "FTS5 rebuild: " + std::to_string(result.succeeded) + " ok, " +
-                     std::to_string(result.failed) + " failed, " + std::to_string(result.skipped) +
-                     " skipped";
-    metaFacade.flush();
-    return result;
-}
-
-boost::asio::awaitable<RepairOperationResult>
-RepairService::generateMissingEmbeddingsAsync(const RepairRequest& req, const ProgressFn& progress,
-                                              std::atomic<bool>* cancelRequested) {
-    RepairOperationResult result;
-    result.operation = "embeddings";
-
-    auto isCanceled = [&]() {
-        return cancelRequested && cancelRequested->load(std::memory_order_relaxed);
-    };
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        co_return result;
-    }
-
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "embeddings";
-        ev.message = "Scanning metadata candidates for missing embeddings";
-        progress(ev);
-    }
-
-    auto candidateScanResult =
-        repair::selectEmbeddingRepairCandidates(*meta, req.includeMime, req.force);
-    if (!candidateScanResult) {
-        result.message = "Failed to query";
-        co_return result;
-    }
-    auto candidateScan = std::move(candidateScanResult.value());
-    const auto& hashes = candidateScan.documentHashes;
-
-    spdlog::info("RepairService::generateMissingEmbeddingsAsync candidates: scanned={} eligible={} "
-                 "eligible_by_mime={} eligible_by_extracted_text={} excluded_samples=[{}] "
-                 "force={} missing_only_query={}",
-                 candidateScan.documentsScanned, hashes.size(), candidateScan.eligibleByMime,
-                 candidateScan.eligibleByExtractedText, candidateScan.excludedSamples.size(),
-                 req.force ? 1 : 0, req.force ? 0 : 1);
-
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "embeddings";
-        ev.total = hashes.size();
-        ev.message =
-            "Found " + std::to_string(hashes.size()) + " eligible documents missing embeddings";
-        progress(ev);
-    }
-
-    result.processed = hashes.size();
-    result.skipped = 0;
-
-    if (hashes.empty()) {
-        result.message = "No eligible documents";
-        co_return result;
-    }
-
-    if (req.dryRun) {
-        result.skipped += hashes.size();
-        result.message = "Would generate embeddings for " + std::to_string(hashes.size()) + " docs";
-        co_return result;
-    }
-
-    std::string modelName = req.embeddingModel;
-    if (modelName.empty() && ctx_.resolvePreferredModel) {
-        try {
-            modelName = ctx_.resolvePreferredModel();
-        } catch (const std::exception& e) {
-            spdlog::debug("RepairService: resolvePreferredModel failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug("RepairService: resolvePreferredModel failed");
-        }
-    }
-    if (modelName.empty() && ctx_.getEmbeddingModelName) {
-        try {
-            modelName = ctx_.getEmbeddingModelName();
-        } catch (const std::exception& e) {
-            spdlog::debug("RepairService: getEmbeddingModelName failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug("RepairService: getEmbeddingModelName failed");
-        }
-    }
-    if (isCanceled()) {
-        result.failed = hashes.size();
-        result.message = "Repair canceled";
-        co_return result;
-    }
-
-    const uint32_t embedCap = TuneAdvisor::embedChannelCapacity();
-    auto embedQ = InternalEventBus::instance().get_or_create_channel<InternalEventBus::EmbedJob>(
-        "embed_jobs", embedCap);
-    const std::size_t batchSize = std::max<std::size_t>(1u, cfg_.maxBatch);
-    boost::asio::steady_timer timer(co_await boost::asio::this_coro::executor);
-
-    for (std::size_t i = 0; i < hashes.size(); i += batchSize) {
-        if (isCanceled()) {
-            result.failed += (hashes.size() - i);
-            result.message = "Repair canceled";
-            co_return result;
-        }
-
-        const std::size_t end = std::min(i + batchSize, hashes.size());
-        std::vector<std::string> batch(hashes.begin() + static_cast<std::ptrdiff_t>(i),
-                                       hashes.begin() + static_cast<std::ptrdiff_t>(end));
-        const std::size_t batchIndex = (i / batchSize) + 1;
-        const std::size_t totalBatches = (hashes.size() + batchSize - 1) / batchSize;
-
-        submitRepairStatusUpdate(ctx_, meta, batch, metadata::RepairStatus::Processing,
-                                 "RepairService::executeBatch/processing");
-
-        InternalEventBus::EmbedJob job{batch,
-                                       static_cast<uint32_t>(batch.size()),
-                                       !req.force,
-                                       modelName,
-                                       std::vector<InternalEventBus::EmbedPreparedDoc>{},
-                                       nullptr};
-        // Batch semantic graph maintenance outside the per-job embedding hot path.
-        // Per-job corpus scans cause large transient heap spikes during repair.
-        job.updateSemanticGraph = false;
-
-        const std::string jobLabel =
-            "embed batch " + std::to_string(batchIndex) + "/" + std::to_string(totalBatches);
-        const bool queued = co_await queueWithBackoff(embedQ, std::move(job), timer, running_,
-                                                      jobLabel, batch.size(), 50, 1000, isCanceled);
-        if (!queued) {
-            submitRepairStatusUpdate(ctx_, meta, batch, metadata::RepairStatus::Pending,
-                                     "RepairService::executeBatch/rollback");
-            InternalEventBus::instance().incEmbedDropped(batch.size());
-            if (isCanceled()) {
-                result.failed += (hashes.size() - i);
-                result.message = "Repair canceled";
-                co_return result;
-            }
-            result.failed += batch.size();
-            continue;
-        }
-
-        InternalEventBus::instance().incEmbedQueued(batch.size());
-        TuningManager::notifyWakeup();
-        result.succeeded += batch.size();
-        result.processed = result.succeeded + result.failed + result.skipped;
-
-        if (progress) {
-            RepairEvent ev;
-            ev.phase = "repairing";
-            ev.operation = "embeddings";
-            ev.processed = std::min(end, hashes.size());
-            ev.total = hashes.size();
-            ev.succeeded = result.succeeded;
-            ev.failed = result.failed;
-            ev.skipped = result.skipped;
-            ev.message = "queued daemon embed batch " + std::to_string(batchIndex) + "/" +
-                         std::to_string(totalBatches);
-            progress(ev);
-        }
-    }
-
-    result.message = "Queued " + std::to_string(result.succeeded) + " docs for embedding";
-    co_return result;
-}
-
-RepairOperationResult RepairService::generateMissingEmbeddings(const RepairRequest& req,
-                                                               const ProgressFn& progress,
-                                                               std::atomic<bool>* cancelRequested) {
-    RepairOperationResult result;
-    result.operation = "embeddings";
-
-    auto isCanceled = [&]() {
-        return cancelRequested && cancelRequested->load(std::memory_order_relaxed);
-    };
-
-    auto meta = ctx_.getMetadataRepo ? ctx_.getMetadataRepo() : nullptr;
-    if (!meta) {
-        result.message = "Metadata not available";
-        return result;
-    }
-
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "embeddings";
-        ev.message = "Scanning metadata candidates for missing embeddings";
-        progress(ev);
-    }
-
-    auto candidateScanResult =
-        repair::selectEmbeddingRepairCandidates(*meta, req.includeMime, req.force);
-    if (!candidateScanResult) {
-        result.message = "Failed to query";
-        return result;
-    }
-    auto candidateScan = std::move(candidateScanResult.value());
-    const auto& hashes = candidateScan.documentHashes;
-
-    spdlog::info("RepairService::generateMissingEmbeddings candidates: scanned={} eligible={} "
-                 "eligible_by_mime={} eligible_by_extracted_text={} excluded_samples=[{}] "
-                 "force={} missing_only_query={}",
-                 candidateScan.documentsScanned, hashes.size(), candidateScan.eligibleByMime,
-                 candidateScan.eligibleByExtractedText, candidateScan.excludedSamples.size(),
-                 req.force ? 1 : 0, req.force ? 0 : 1);
-
-    if (progress) {
-        RepairEvent ev;
-        ev.phase = "repairing";
-        ev.operation = "embeddings";
-        ev.total = hashes.size();
-        ev.message =
-            "Found " + std::to_string(hashes.size()) + " eligible documents missing embeddings";
-        progress(ev);
-    }
-
-    size_t skippedExisting = 0;
-
-    result.processed = hashes.size() + skippedExisting;
-    result.skipped = skippedExisting;
-
-    if (hashes.empty()) {
-        if (skippedExisting > 0) {
-            result.message = "All " + std::to_string(skippedExisting) +
-                             " eligible documents already have embeddings";
-        } else {
-            result.message = "No eligible documents";
-        }
-        return result;
-    }
-
-    if (req.dryRun) {
-        result.skipped += hashes.size();
-        result.message = "Would generate embeddings for " + std::to_string(hashes.size()) + " docs";
-        if (skippedExisting > 0) {
-            result.message +=
-                " (" + std::to_string(skippedExisting) + " already embedded, skipped)";
-        }
-        return result;
-    }
-
-    std::string modelName = req.embeddingModel;
-    if (modelName.empty() && ctx_.resolvePreferredModel) {
-        try {
-            modelName = ctx_.resolvePreferredModel();
-        } catch (const std::exception& e) {
-            spdlog::debug("RepairService: resolvePreferredModel failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug("RepairService: resolvePreferredModel failed");
-        }
-    }
-    if (modelName.empty() && ctx_.getEmbeddingModelName) {
-        try {
-            modelName = ctx_.getEmbeddingModelName();
-        } catch (const std::exception& e) {
-            spdlog::debug("RepairService: getEmbeddingModelName failed: {}", e.what());
-        } catch (...) {
-            spdlog::debug("RepairService: getEmbeddingModelName failed");
-        }
-    }
-    if (req.foreground && modelName.empty()) {
-        result.failed = hashes.size();
-        result.message = "No embedding model configured";
-        return result;
-    }
-    if (isCanceled()) {
-        result.failed = hashes.size();
-        result.message = "Repair canceled";
-        return result;
-    }
-
-    const uint32_t embedCap = TuneAdvisor::embedChannelCapacity();
-    auto embedQ = InternalEventBus::instance().get_or_create_channel<InternalEventBus::EmbedJob>(
-        "embed_jobs", embedCap);
-
-    const std::size_t batchSize = std::max<std::size_t>(1u, cfg_.maxBatch);
-    const std::size_t totalDocs = hashes.size() + skippedExisting;
-    const std::size_t totalBatches = (hashes.size() + batchSize - 1) / batchSize;
-    std::size_t completedDocs = 0;
-    std::size_t failedDocs = 0;
-    std::size_t skippedDocs = skippedExisting;
-
-    for (std::size_t i = 0; i < hashes.size(); i += batchSize) {
-        if (isCanceled()) {
-            result.failed += (hashes.size() - i);
-            result.message = "Repair canceled";
-            return result;
-        }
-
-        const std::size_t end = std::min(i + batchSize, hashes.size());
-        std::vector<std::string> batch(hashes.begin() + i, hashes.begin() + end);
-        const std::size_t batchIndex = i / batchSize;
-
-        submitRepairStatusUpdate(ctx_, meta, batch, metadata::RepairStatus::Processing,
-                                 "RepairService::executeBatchMonitored/processing");
-
-        std::shared_ptr<InternalEventBus::EmbedJobMonitor> monitor;
-        if (req.foreground) {
-            monitor = std::make_shared<InternalEventBus::EmbedJobMonitor>();
-            monitor->totalDocs = batch.size();
-            monitor->phase = "queued";
-            monitor->detail = "queued daemon embed batch";
-        }
-
-        InternalEventBus::EmbedJob job{batch,
-                                       static_cast<uint32_t>(batch.size()),
-                                       !req.force,
-                                       modelName,
-                                       std::vector<InternalEventBus::EmbedPreparedDoc>{},
-                                       monitor};
-        // Batch semantic graph maintenance outside the per-job embedding hot path.
-        // Per-job corpus scans cause large transient heap spikes during repair.
-        job.updateSemanticGraph = false;
-
-        int retries = 0;
-        bool pushed = false;
-        while (!pushed && retries < 20) {
-            pushed = embedQ->try_push(job);
-            if (!pushed) {
-                std::this_thread::sleep_for(
-                    std::chrono::milliseconds(50 * (1 << std::min(retries, 5))));
-                retries++;
-            }
-        }
-
-        if (!pushed) {
-            result.failed += batch.size();
-            submitRepairStatusUpdate(ctx_, meta, batch, metadata::RepairStatus::Pending,
-                                     "RepairService::executeBatchMonitored/rollback");
-            InternalEventBus::instance().incEmbedDropped();
-            continue;
-        }
-
-        InternalEventBus::instance().incEmbedQueued();
-        TuningManager::notifyWakeup();
-
-        if (!req.foreground) {
-            result.succeeded += batch.size();
-            if (progress) {
-                RepairEvent ev;
-                ev.phase = "repairing";
-                ev.operation = "embeddings";
-                ev.processed = std::min(i + batchSize, hashes.size());
-                ev.total = hashes.size();
-                ev.succeeded = result.succeeded;
-                ev.failed = result.failed;
-                ev.skipped = result.skipped;
-                progress(ev);
-            }
-            continue;
-        }
-
-        std::size_t lastReportedProcessed = static_cast<std::size_t>(-1);
-        auto lastReport = std::chrono::steady_clock::time_point{};
-        while (true) {
-            if (isCanceled()) {
-                if (monitor) {
-                    monitor->cancelRequested.store(true, std::memory_order_relaxed);
-                    monitor->cv.notify_all();
-                }
-                result.message = "Repair canceled";
-                return result;
-            }
-
-            std::size_t batchCompleted = 0;
-            std::size_t batchFailed = 0;
-            std::size_t batchSkipped = 0;
-            std::size_t batchProcessed = 0;
-            std::string batchPhase = "queued";
-            std::string batchDetail;
-            bool batchDone = false;
-            {
-                std::unique_lock<std::mutex> lk(monitor->mutex);
-                monitor->cv.wait_for(lk, std::chrono::milliseconds(250));
-                batchCompleted = static_cast<std::size_t>(monitor->succeededDocs);
-                batchFailed = static_cast<std::size_t>(monitor->failedDocs);
-                batchSkipped = static_cast<std::size_t>(monitor->skippedDocs);
-                batchProcessed = static_cast<std::size_t>(monitor->processedDocs);
-                batchPhase = monitor->phase;
-                batchDetail = monitor->detail;
-                batchDone = monitor->done;
-            }
-
-            const auto now = std::chrono::steady_clock::now();
-            if (progress && (batchProcessed != lastReportedProcessed ||
-                             lastReport.time_since_epoch().count() == 0 ||
-                             (now - lastReport) >= std::chrono::seconds(1))) {
-                RepairEvent ev;
-                ev.phase = "repairing";
-                ev.operation = "embeddings";
-                ev.processed = completedDocs + failedDocs + skippedDocs + batchProcessed;
-                ev.total = totalDocs;
-                ev.succeeded = completedDocs + batchCompleted;
-                ev.failed = failedDocs + batchFailed;
-                ev.skipped = skippedDocs + batchSkipped;
-                if (batchDone) {
-                    ev.message = "completed daemon embed batch " + std::to_string(batchIndex + 1) +
-                                 "/" + std::to_string(totalBatches);
-                } else {
-                    ev.message = "daemon embed batch " + std::to_string(batchIndex + 1) + "/" +
-                                 std::to_string(totalBatches) + " phase=" + batchPhase;
-                    if (!batchDetail.empty()) {
-                        ev.message += " " + batchDetail;
-                    }
-                }
-                progress(ev);
-                lastReportedProcessed = batchProcessed;
-                lastReport = now;
-            }
-
-            if (batchDone) {
-                completedDocs += batchCompleted;
-                failedDocs += batchFailed;
-                skippedDocs += batchSkipped;
-                result.processed = completedDocs + failedDocs + skippedDocs;
-                result.succeeded = completedDocs;
-                result.failed = failedDocs;
-                result.skipped = skippedDocs;
-                break;
-            }
-
-            std::this_thread::sleep_for(std::chrono::milliseconds(250));
-        }
-    }
-
-    if (req.foreground) {
-        // No explicit finalize needed — the coordinator manages the bulk window.
-        if (req.repairTopology && result.succeeded > 0 && ctx_.rebuildSemanticNeighborGraph) {
-            auto semanticResult =
-                ctx_.rebuildSemanticNeighborGraph("repair_service.embedding_batch", modelName);
-            if (!semanticResult) {
-                spdlog::warn("RepairService: deferred semantic neighbor rebuild failed: {}",
-                             semanticResult.error().message);
-            } else if (progress) {
-                RepairEvent ev;
-                ev.phase = "repairing";
-                ev.operation = "embeddings";
-                ev.processed = result.processed;
-                ev.total = totalDocs;
-                ev.succeeded = result.succeeded;
-                ev.failed = result.failed;
-                ev.skipped = result.skipped;
-                ev.message = "rebuilt semantic neighbor graph edges=" +
-                             std::to_string(semanticResult.value());
-                progress(ev);
-            }
-        }
-        result.message = "Generated " + std::to_string(result.succeeded) + " embeddings";
-        if (result.skipped > 0 || result.failed > 0) {
-            result.message += ", skipped=" + std::to_string(result.skipped) +
-                              ", failed=" + std::to_string(result.failed);
-        }
-        if (state_) {
-            state_->stats.repairEmbeddingsGenerated.store(result.succeeded,
-                                                          std::memory_order_relaxed);
-            state_->stats.repairEmbeddingsSkipped.store(result.skipped, std::memory_order_relaxed);
-            state_->stats.repairFailedOperations.store(result.failed, std::memory_order_relaxed);
-        }
-    } else {
-        result.message = "Queued " + std::to_string(result.succeeded) + " docs for embedding";
-        if (result.skipped > 0) {
-            result.message += " (" + std::to_string(result.skipped) + " already embedded, skipped)";
-        }
-    }
-    return result;
-}
-
-RepairOperationResult RepairService::optimizeDatabase(bool dryRun, bool verbose,
-                                                      ProgressFn progress) {
-    RepairOperationResult result;
-    result.operation = "optimize";
-
-    namespace fs = std::filesystem;
-    fs::path dbPath = cfg_.dataDir / "yams.db";
-    fs::path vecDbPath = cfg_.dataDir / "vectors.db";
-
-    int dbCount = 0;
-    if (fs::exists(dbPath))
-        dbCount++;
-    if (fs::exists(vecDbPath))
-        dbCount++;
-
-    if (dbCount == 0) {
-        result.message = "No databases found";
-        return result;
-    }
-
-    result.processed = dbCount;
-
-    if (dryRun) {
-        result.skipped = dbCount;
-        result.message = "Would optimize " + std::to_string(dbCount) + " database(s)";
-        return result;
-    }
-
-    int succeeded = 0;
-    std::string failMsg;
-
-    // Helper: run live-safe SQLite maintenance for a single database.
-    //
-    // Avoid VACUUM here. `repair` runs inside the daemon while the same process owns long-lived
-    // metadata/vector connections. A full VACUUM needs heavyweight locks and can create a large
-    // transient database copy; on large stores this can stall repair streaming or exhaust disk/RSS,
-    // surfacing to the CLI as `[ipc:eof]` if the daemon disconnects. Keep daemon repair maintenance
-    // bounded: checkpoint opportunistically and let SQLite's PRAGMA optimize choose cheap work.
-    auto optimizeOne = [&](const fs::path& path, const char* label) {
-        sqlite3* db = nullptr;
-        if (sqlite3_open_v2(path.string().c_str(), &db, SQLITE_OPEN_READWRITE, nullptr) !=
-            SQLITE_OK) {
-            failMsg += std::string(label) + ": failed to open; ";
-            if (db)
-                sqlite3_close(db);
-            return false;
-        }
-        sqlite3_busy_timeout(db, 5000);
-
-        auto emit = [&](std::string phase, std::string message) {
-            if (!progress)
-                return;
-            RepairEvent ev;
-            ev.operation = "optimize";
-            ev.phase = std::move(phase);
-            ev.processed = result.processed;
-            ev.succeeded = succeeded;
-            ev.failed = result.failed;
-            ev.message = std::string(label) + ": " + message;
-            progress(ev);
-        };
-
-        emit("repairing", "checkpointing WAL");
-        int walLog = 0, walCkpt = 0;
-        sqlite3_wal_checkpoint_v2(db, nullptr, SQLITE_CHECKPOINT_PASSIVE, &walLog, &walCkpt);
-        spdlog::info("[Optimize] {} WAL checkpoint: log={} checkpointed={}", label, walLog,
-                     walCkpt);
-
-        auto exec = [&](const char* sql) -> bool {
-            char* errMsg = nullptr;
-            const int rc = sqlite3_exec(db, sql, nullptr, nullptr, &errMsg);
-            if (rc == SQLITE_OK)
-                return true;
-            std::string error = errMsg ? errMsg : sqlite3_errmsg(db);
-            sqlite3_free(errMsg);
-            failMsg += std::string(label) + ": " + sql + " failed: " + error + "; ";
-            return false;
-        };
-
-        emit("repairing", "running PRAGMA optimize");
-        bool ok = exec("PRAGMA analysis_limit=1000") && exec("PRAGMA optimize");
-        if (ok) {
-            spdlog::info("[Optimize] {} PRAGMA optimize completed", label);
-            emit("repairing", "optimized");
-        }
-        sqlite3_close(db);
-        return ok;
-    };
-
-    if (fs::exists(dbPath) && optimizeOne(dbPath, "yams.db")) {
-        succeeded++;
-    }
-    if (fs::exists(vecDbPath) && optimizeOne(vecDbPath, "vectors.db")) {
-        succeeded++;
-    }
-
-    result.succeeded = succeeded;
-    result.failed = dbCount - succeeded;
-
-    if (result.failed == 0) {
-        result.message = "All databases optimized";
-    } else {
-        result.message = "Optimize partially failed: " + failMsg;
-    }
-
-    return result;
-}
 
 } // namespace yams::daemon
