+++ src/daemon/meson.build
@@ -136,6 +136,21 @@ yams_daemon_lib = static_library('yams_daemon',
     'components/VectorIndexCoordinator.cpp',
     'components/repair/repair_health_probe.cpp',
     'components/repair/repair_plan_builder.cpp',
+    'components/repair/repair_operation_registry.cpp',
+    'components/repair/repair_operation_support.cpp',
+    'components/repair/ops/stuck_documents.cpp',
+    'components/repair/ops/orphaned_metadata.cpp',
+    'components/repair/ops/mime_types.cpp',
+    'components/repair/ops/downloads.cpp',
+    'components/repair/ops/path_tree.cpp',
+    'components/repair/ops/semantic_dedupe.cpp',
+    'components/repair/ops/orphaned_chunks.cpp',
+    'components/repair/ops/block_references.cpp',
+    'components/repair/ops/knowledge_graph.cpp',
+    'components/repair/ops/fts5_index.cpp',
+    'components/repair/ops/missing_embeddings.cpp',
+    'components/repair/ops/topology_artifacts.cpp',
+    'components/repair/ops/optimize_database.cpp',
     'components/EntityGraphService.cpp',
     'components/gliner_query_extractor.cpp',
     'components/SocketServer.cpp',
